refactor: 将`simulatedPlayers`实例移动到`SimPlayerList.ts` 不导出`SimulatedPlayerList`
